+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## Unreleased
+
+* Changed namespace from `Skyzyx\UtilityPack` → `SimplePie\UtilityPack\Util`.
+* Updated a bunch of linting rules.
+
 ## 1.0.0 - 2017-05-27
 
 * Extracted from code that I use all the time.